+++ apps/frontend/src/modules/points-of-interest/slices/actions.ts
@@ -1,14 +1,19 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 
-import { type APIResponse, type AsyncThunkConfig } from "~/libs/types/types.js";
 import {
-	type PointsOfInterestGetByIdResponseDto,
-	type PointsOfInterestPaginatedResponseDto,
-	type PointsOfInterestQueryRequest,
-	type PointsOfInterestRequestDto,
-	type PointsOfInterestResponseDto,
-} from "~/modules/points-of-interest/points-of-interest.js";
+	type APIResponse,
+	type AsyncThunkConfig,
+	type PaginationMeta,
+} from "~/libs/types/types.js";
 
+import { PointOfInterestNotification } from "../libs/enums/enums.js";
+import {
+	type PatchActionPayload,
+	type PointsOfInterestCreateRequestDto,
+	type PointsOfInterestGetAllItemResponseDto,
+	type PointsOfInterestGetAllQuery,
+	type PointsOfInterestGetByIdResponseDto,
+} from "../libs/types/types.js";
 import { name as poiDetailsSliceName } from "./points-of-interest-details.slice.js";
 import { name as poiSliceName } from "./points-of-interest.slice.js";
 
@@ -23,8 +28,8 @@ const getById = createAsyncThunk<
 });
 
 const create = createAsyncThunk<
-	APIResponse<PointsOfInterestResponseDto>,
-	PointsOfInterestRequestDto,
+	APIResponse<PointsOfInterestGetByIdResponseDto>,
+	PointsOfInterestCreateRequestDto,
 	AsyncThunkConfig
 >(`${poiSliceName}/create`, async (payload, { extra }) => {
 	const { pointOfInterestApi, toastNotifier } = extra;
@@ -35,16 +40,29 @@ const create = createAsyncThunk<
 	return pointOfInterest;
 });
 
-const findPaginated = createAsyncThunk<
-	APIResponse<PointsOfInterestPaginatedResponseDto>,
-	PointsOfInterestQueryRequest,
+const findAll = createAsyncThunk<
+	APIResponse<PointsOfInterestGetAllItemResponseDto[], PaginationMeta>,
+	PointsOfInterestGetAllQuery,
 	AsyncThunkConfig
->(`${poiSliceName}/find-paginated-items`, async (payload, { extra }) => {
+>(`${poiSliceName}/find-all`, async (payload, { extra }) => {
 	const { pointOfInterestApi } = extra;
 
-	const pointsOfInterest = await pointOfInterestApi.findPaginated(payload);
+	const { data, meta } = await pointOfInterestApi.findAll(payload);
 
-	return pointsOfInterest;
+	return { data, meta };
 });
 
-export { create, findPaginated, getById };
+const patch = createAsyncThunk<
+	APIResponse<PointsOfInterestGetByIdResponseDto>,
+	PatchActionPayload,
+	AsyncThunkConfig
+>(`${poiSliceName}/patch-point-of-interest`, async (payload, { extra }) => {
+	const { pointOfInterestApi, toastNotifier } = extra;
+	const patchResult = await pointOfInterestApi.patch(payload);
+
+	toastNotifier.showSuccess(PointOfInterestNotification.UPDATED);
+
+	return patchResult;
+});
+
+export { create, findAll, getById, patch };