+++ apps/frontend/src/modules/points-of-interest/slices/points-of-interest-details.slice.ts
@@ -4,15 +4,17 @@ import { DataStatus } from "~/libs/enums/enums.js";
 import { type ValueOf } from "~/libs/types/types.js";
 import { type PointsOfInterestGetByIdResponseDto } from "~/modules/points-of-interest/points-of-interest.js";
 
-import { getById } from "./actions.js";
+import { getById, patch } from "./actions.js";
 
 type State = {
 	dataStatus: ValueOf<typeof DataStatus>;
+	patchStatus: ValueOf<typeof DataStatus>;
 	pointsOfInterestDetails: null | PointsOfInterestGetByIdResponseDto;
 };
 
 const initialState: State = {
 	dataStatus: DataStatus.IDLE,
+	patchStatus: DataStatus.IDLE,
 	pointsOfInterestDetails: null,
 };
 
@@ -27,6 +29,16 @@ const { actions, name, reducer } = createSlice({
 		});
 		builder.addCase(getById.rejected, (state) => {
 			state.dataStatus = DataStatus.REJECTED;
+		});
+		builder.addCase(patch.pending, (state) => {
+			state.patchStatus = DataStatus.PENDING;
+		});
+		builder.addCase(patch.fulfilled, (state, action) => {
+			state.pointsOfInterestDetails = action.payload.data;
+			state.patchStatus = DataStatus.FULFILLED;
+		});
+		builder.addCase(patch.rejected, (state) => {
+			state.patchStatus = DataStatus.REJECTED;
 		});
 	},
 	initialState,