+++ packages/shared/src/modules/points-of-interest/libs/types/points-of-interest-paginated-response-dto.type.ts
@@ -1,9 +1,0 @@
-import { type PaginationMeta } from "../../../../libs/types/pagination-meta.type.js";
-import { type PointsOfInterestItemDto } from "./points-of-interest-item-dto.type.js";
-
-type PointsOfInterestPaginatedResponseDto = {
-	data: PointsOfInterestItemDto[];
-	meta: PaginationMeta;
-};
-
-export { type PointsOfInterestPaginatedResponseDto };